feat: Refactor UtilityCog and introduce core event handling
- Removed the existing UtilityCog implementation and replaced it with a new structure that includes utility commands and an admin subgroup for enhanced functionality.
- Added a new CoreEventsCog to handle core Discord events, improving logging and monitoring capabilities.
- Introduced a TasksCog for background tasks, including periodic checks for CISA KEV feed and stats reporting to a web application.
- Enhanced the help command to provide detailed usage information for commands and subcommands.
- Implemented uptime functionality to display the bot's running time.
- Updated the bot's initialization process to accommodate the new cog structure and improve overall organization.
